+++ src/MapPoint.hpp
@@ -16,6 +16,7 @@ namespace SLAM_demo {
 
 // forward declarations
 class Frame;
+class Map;
 
 /**
  * @class MapPoint
@@ -30,14 +31,7 @@ public: // public members
      * @param[in] X3D 3D world coordinate as a \f$3 \times 1\f$ matrix.
      * @param[in] nIdxFrm Index of the frame that observed the point.
      */
-    MapPoint(const cv::Mat& X3D, int nIdxFrm);
-    /** (OBSOLETE)
-     * @brief Construct map point with input coordinate and descriptor.
-     * @param[in] X3D     3D world coordinate as a \f$3 \times 1\f$ matrix.
-     * @param[in] desc    Descriptor of the map point.
-     * @param[in] nIdxFrm Index of the frame that observed the point.
-     */
-    MapPoint(const cv::Mat& X3D, const cv::Mat& desc, int nIdxFrm);
+    MapPoint(const std::shared_ptr<Map>& pMap, const cv::Mat& X3D);
     /// Get 3D world coordinate.
     cv::Mat getX3D() const { return mX3D; }
     /// Get most distinctive descriptor of the point.
@@ -50,8 +44,8 @@ public: // public members
     std::vector<std::shared_ptr<Frame>> getRelatedFrames() const;
     /// Get index of the latest frame that observed the point.
     unsigned getIdxLastObsFrm() const { return mnIdxLastObsFrm; }
-    /// Get match-to-observation ratio.
-    float getMatch2ObsRatio() const;
+    /// Get observation-to-visible ratio.
+    float getObs2VisibleRatio() const;
     /// Update 3D world coordinate of the point.
     void setX3D(const cv::Mat& X3D) { mX3D = X3D.clone(); }
     /// Update descriptor of the point.
@@ -59,9 +53,7 @@ public: // public members
     /// Update index of the latest frame that observed the point.
     void setIdxLastObsFrm(int idx) { mnIdxLastObsFrm = idx; }
     /// Update count of being observed by input frames.
-    void addCntObs(int n = 1);
-    /// Update count of being matched by other points.
-    void addCntMatches(int n = 1);
+    void addCntVisible(int n = 1);
     /// Check whether a map point is an outlier.
     bool isOutlier() const { return mbOutlier; }
     /// Set outlier flag.
@@ -72,8 +64,14 @@ public: // public members
      * @param[in] nIdxKpt Index of the keypoint extracted from the frame.
      */
     void addObservation(const std::shared_ptr<Frame>& pFrame, int nIdxKpt);
+    void removeObservation(const std::shared_ptr<Frame>& pFrame);
+    /// Update the best descriptor for the map point.
     void updateDescriptor();
+    /// Whether the map point is observed by a frame or not.
+    bool isObservedBy(const std::shared_ptr<Frame>& pFrame) const;
 private: // private data
+    // Pointer to the map.
+    std::shared_ptr<Map> mpMap;
     /// Inhomogeneous 3D world coordinate of the point.
     cv::Mat mX3D;
     /// Observation data (pointer to frame, index of the observed keypoint).
@@ -83,14 +81,17 @@ private: // private data
     /// Index of the latest frame by which the point is observed.
     unsigned mnIdxLastObsFrm;
     /// Number of times by which the point is observed by input frames.
-    int mnCntObs;
+    int mnCntVisible;
     /** 
      * @brief Number of times by which the point is matched with another 
      *        point (2D/3D) after the system is initialized.
      */
-    int mnCntMatches;
+    int mnCntObs;
     /// Whether the map point is an outlier.
     bool mbOutlier;
+private: // private members
+    /// Update count of being matched by other points.
+    void addCntObs(int n = 1);
 };
 
 } // namespace SLAM_demo